refactor: Rename fake to mock
