+++ src/qibocal/auto/platforms/mock/calibration.json
