+++ src/qibocal/auto/platforms/mock/parameters.json
