+++ src/qibocal/auto/runcard.py
@@ -32,7 +32,7 @@ class Runcard:
     """
     backend: str = "qibolab"
     """Qibo backend."""
-    platform: str = "fake"
+    platform: str = "mock"
     """Qibolab platform."""
     update: bool = True
