+++ tests/runcards/dummy_action.yml
@@ -1,5 +1,5 @@
 backend: qibolab
-platform: fake
+platform: mock
 
 targets: [0,1]
