+++ tests/runcards/dummy_compare.yml
@@ -1,5 +1,5 @@
 backend: qibolab
-platform: fake
+platform: mock
 
 targets: [0]
