+++ tests/runcards/protocols.yml
@@ -1,4 +1,4 @@
-platform: [fake]
+platform: [mock]
 
 targets: [0,1]
