+++ tests/test_output.py
@@ -28,8 +28,8 @@ ACTION = Action(**action)
 # TODO: this is essentially a proto `qq run` invocation, it should be simplified as
 # much as possible in the library, and made available in conftest
 @pytest.fixture
-def fake_output(tmp_path: Path) -> tuple[Output, Path]:
-    backend = construct_backend(backend="qibolab", platform="fake")
+def mock_output(tmp_path: Path) -> tuple[Output, Path]:
+    backend = construct_backend(backend="qibolab", platform="mock")
     platform: Platform = backend.platform
     meta = Metadata.generate(tmp_path.name, backend)
     output = Output(History(), meta, platform)
@@ -47,9 +47,9 @@ def fake_output(tmp_path: Path) -> tuple[Output, Path]:
     return output, tmp_path
 
 
-def test_output_process(fake_output: tuple[Output, Path]):
+def test_output_process(mock_output: tuple[Output, Path]):
     """Create method of Executor."""
-    output, path = fake_output
+    output, path = mock_output
     # perform fit
     output.process(path, mode=ExecutionMode.FIT)
